feat: 타입으로 바우처 조회 구현 및 테스트 코드 작성

+++ src/main/java/org/promgrammers/springbootbasic/domain/voucher/service/VoucherService.java
@@ -8,6 +8,7 @@ import org.promgrammers.springbootbasic.domain.voucher.dto.request.UpdateVoucher
 import org.promgrammers.springbootbasic.domain.voucher.dto.response.VoucherListResponse;
 import org.promgrammers.springbootbasic.domain.voucher.dto.response.VoucherResponse;
 import org.promgrammers.springbootbasic.domain.voucher.model.Voucher;
+import org.promgrammers.springbootbasic.domain.voucher.model.VoucherType;
 import org.promgrammers.springbootbasic.domain.voucher.repository.VoucherRepository;
 import org.promgrammers.springbootbasic.exception.BusinessException;
 import org.slf4j.Logger;
@@ -56,6 +57,7 @@ public class VoucherService {
         }
 
         wallet.addVoucher(voucher);
+        System.out.println(wallet.getVouchers().toString());
 
         voucherRepository.assignVoucherToCustomer(customer.getCustomerId(), voucher.getVoucherId());
     }
@@ -81,6 +83,21 @@ public class VoucherService {
                 .orElseThrow(() -> new BusinessException(NOT_FOUND_VOUCHER));
 
         return new VoucherResponse(voucher);
+    }
+
+    @Transactional(readOnly = true)
+    public VoucherListResponse findByType(VoucherType voucherType) {
+        List<Voucher> voucherList = voucherRepository.findByType(voucherType);
+
+        if (voucherList == null || voucherList.isEmpty()) {
+            throw new BusinessException(NOT_FOUND_VOUCHER);
+        }
+
+        List<VoucherResponse> voucherResponseList = voucherList.stream()
+                .map(VoucherResponse::new)
+                .toList();
+
+        return new VoucherListResponse(voucherResponseList);
     }
 
     @Transactional(readOnly = true)
@@ -131,6 +148,11 @@ public class VoucherService {
 
     @Transactional
     public void deleteAll() {
+        List<Voucher> voucherList = voucherRepository.findAll();
+
+        if (voucherList == null || voucherList.isEmpty()) {
+            throw new BusinessException(NOT_FOUND_VOUCHER);
+        }
         voucherRepository.deleteAll();
     }
